requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
Django==3.1.6
djangorestframework==3.12.2
idna==2.10
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.18.0
joblib==0.14.1
mysqlclient==1.4.6
nltk==3.5
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tqdm==4.56.2
traitlets==4.3.3
urllib3==1.26.3
wcwidth==0.2.5
